requirements.txt
# environment: Python 3.13.5

# Core
torch==2.8.0+cu128
numpy==2.1.3
scipy==1.15.3
pandas==2.2.3
tqdm==4.67.1
click==8.1.8
dill==0.4.0
hydra-core==1.3.2
omegaconf==2.3.0
wandb==0.25.1
einops==0.8.1
diffusers==0.37.1
numba==0.61.0

# Data / storage
zarr==2.18.7
numcodecs==0.15.1
h5py==3.12.1
imageio==2.37.0

# Computer vision / simulation
opencv-python==4.12.0.88
scikit-image==0.25.0
shapely==2.1.2
pygame==2.6.1
pymunk==7.2.0
matplotlib==3.10.0
gymnasium==1.2.3
